Adds SwiftLint, as build phase
Currently configured in such a way that following the next, "compliance" commit, the highest severity message will be "Warning".
A number of rules are excluded in `swiftlint.yml` -- generally those which could be considered as explicit, discretionary choices of the programmer or as making code more explicit, or those which would require a more significant refactor, which could be done at a later date. Obviously, these can and should be reviewed.