Bug fix (#463): When tapping on non-editable text, player shortcut events get swallowed
This issue happened, for example, if the user tapped in the description area underneath the player, which caused the labels to become first responders. However, these labels cannot be edited, unlike the search bar on the left pane of the app, causing frustrating behaviour due to keyboard shortcuts sometimes working and sometimes not for no apparent reason.