Fixed Timestamp preview blurriness in Retina screens
I believe the issues is that on buildUI(), we don't yet have a value for the screen's backing scale factor, so we default to "1". By updating it on every NSEvent, it's more expensive, but at this point I don't think it's worth the point of doing a second "setup" phase for the timestamp preview layer.