Bug fix: When de-selecting an annotation with the mouse, we no longer show the time preview
This code was causing issues when dragging annotations around the timeline to move them, our out of the timeline to delete them/remove them. The code we remove caused the preview layer to be shown even though the user wasn't scrubbing the mouse over the timeline's bar. Now, when the user returns to scrubbing the timeline, the existing code will re-animate in the time preview layer.